requirements.txt
yfinance==0.2.32
pandas==2.1.4
numpy==1.26.3
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
scipy==1.11.4
scikit-learn==1.3.2
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
tabulate==0.9.0
openpyxl==3.1.0
